move connect metamask button onto bridge page
