+++ components/bridge/Deposit.tsx
@@ -10,6 +10,9 @@ import { useCENNZApi } from "../../providers/CENNZApiProvider";
 import TxModal from "./TxModal";
 import TokenPicker from "../shared/TokenPicker";
 import CENNZnetAccountPicker from "./CENNZnetAccountPicker";
+import ErrorModal from "./ErrorModal";
+
+const ETH_CHAIN_ID = process.env.NEXT_PUBLIC_ETH_CHAIN_ID;
 
 const Deposit: React.FC<{}> = () => {
 	const [customAddress, setCustomAddress] = useState(false);
@@ -20,14 +23,46 @@ const Deposit: React.FC<{}> = () => {
 		name: "",
 	});
 	const [modalOpen, setModalOpen] = useState(false);
+	const [errorModalOpen, setErrorModalOpen] = useState(false);
+	const [modalState, setModalState] = useState("");
 	const [modal, setModal] = useState({
 		state: "",
 		text: "",
 		hash: "",
 	});
 	const [tokenBalance, setTokenBalance] = useState<Number>();
-	const { Contracts, Signer, Account }: any = useBlockchain();
+	const { Contracts, Signer, Account, initBlockchain }: any = useBlockchain();
 	const { api }: any = useCENNZApi();
+
+	const connectMetamask = async () => {
+		const { ethereum } = window as any;
+		try {
+			const accounts = await ethereum.request({
+				method: "eth_requestAccounts",
+			});
+			const ethChainId = await ethereum.request({ method: "eth_chainId" });
+
+			if (ETH_CHAIN_ID === "1" && ethChainId !== "0x1") {
+				await ethereum.request({
+					method: "wallet_switchEthereumChain",
+					params: [{ chainId: "0x1" }],
+				});
+				window.location.reload();
+			} else if (ETH_CHAIN_ID === "42" && ethChainId !== "0x2a") {
+				await ethereum.request({
+					method: "wallet_switchEthereumChain",
+					params: [{ chainId: "0x2a" }],
+				});
+				window.location.reload();
+			}
+
+			initBlockchain(ethereum, accounts);
+		} catch (err) {
+			console.log(err.message);
+			setModalState("noMetamask");
+			setErrorModalOpen(true);
+		}
+	};
 
 	//Check MetaMask account has enough tokens to deposit
 	useEffect(() => {
@@ -111,6 +146,9 @@ const Deposit: React.FC<{}> = () => {
 					etherscanHash={modal.hash}
 					resetModal={resetModal}
 				/>
+			)}
+			{errorModalOpen && (
+				<ErrorModal setModalOpen={setErrorModalOpen} modalState={modalState} />
 			)}
 			<Box
 				component="form"
@@ -200,27 +238,49 @@ const Deposit: React.FC<{}> = () => {
 						</Button>
 					</>
 				)}
-				<Button
-					sx={{
-						fontFamily: "Teko",
-						fontWeight: "bold",
-						fontSize: "21px",
-						lineHeight: "124%",
-						color: "#1130FF",
-						mt: "30px",
-						mb: "50px",
-					}}
-					disabled={
-						amount && token && selectedAccount && Number(amount) <= tokenBalance
-							? false
-							: true
-					}
-					size="large"
-					variant="outlined"
-					onClick={deposit}
-				>
-					Deposit
-				</Button>
+				{Account ? (
+					<Button
+						sx={{
+							fontFamily: "Teko",
+							fontWeight: "bold",
+							fontSize: "21px",
+							lineHeight: "124%",
+							color: "#1130FF",
+							mt: "30px",
+							mb: "50px",
+						}}
+						disabled={
+							amount &&
+							token &&
+							selectedAccount &&
+							Number(amount) <= tokenBalance
+								? false
+								: true
+						}
+						size="large"
+						variant="outlined"
+						onClick={deposit}
+					>
+						confirm
+					</Button>
+				) : (
+					<Button
+						sx={{
+							fontFamily: "Teko",
+							fontWeight: "bold",
+							fontSize: "21px",
+							lineHeight: "124%",
+							color: "#1130FF",
+							mt: "30px",
+							mb: "50px",
+						}}
+						size="large"
+						variant="outlined"
+						onClick={connectMetamask}
+					>
+						connect metamask
+					</Button>
+				)}
 			</Box>
 		</>
 	);
