+++ pages/_app.tsx
@@ -15,6 +15,7 @@ import Web3AccountsProvider from "../providers/Web3AccountsProvider";
 import Switch from "../components/AppSwitch";
 import Wallet from "../components/Wallet";
 import SupportedAssetsProvider from "../providers/SupportedAssetsProvider";
+import BlockchainProvider from "../providers/BlockchainProvider";
 
 function MyApp({ Component, pageProps }: AppProps) {
 	const router = useRouter();
@@ -39,24 +40,26 @@ function MyApp({ Component, pageProps }: AppProps) {
 						<Web3AccountsProvider>
 							<SupportedAssetsProvider>
 								<SupportedWalletProvider>
-									<Wallet />
-									<Box
-										onClick={() => setLocation("index")}
-										sx={{ cursor: "pointer" }}
-									>
-										<img
-											src="/cennznet-header.png"
-											alt="CENNZnet header"
-											style={{
-												width: isBrowser || isTablet ? "90px" : "45px",
-												position: "absolute",
-												top: "5%",
-												left: "6%",
-											}}
-										/>
-									</Box>
-									<Switch location={location} setLocation={setLocation} />
-									<Component {...pageProps} />
+									<BlockchainProvider>
+										<Wallet />
+										<Box
+											onClick={() => setLocation("index")}
+											sx={{ cursor: "pointer" }}
+										>
+											<img
+												src="/cennznet-header.png"
+												alt="CENNZnet header"
+												style={{
+													width: isBrowser || isTablet ? "90px" : "45px",
+													position: "absolute",
+													top: "5%",
+													left: "6%",
+												}}
+											/>
+										</Box>
+										<Switch location={location} setLocation={setLocation} />
+										<Component {...pageProps} />
+									</BlockchainProvider>
 								</SupportedWalletProvider>
 							</SupportedAssetsProvider>
 						</Web3AccountsProvider>
