+++ pages/bridge/index.tsx
@@ -1,15 +1,15 @@
 import React, { useEffect, useState } from "react";
-import Connect from "./connect";
-import Emery from "./emery";
-import BlockchainProvider from "../../providers/BlockchainProvider";
-import { useWallet } from "../../providers/SupportedWalletProvider";
+import { Typography } from "@mui/material";
+import { Frame, Heading, SmallText } from "../../theme/StyledComponents";
+import { useBlockchain } from "../../providers/BlockchainProvider";
 import { useCENNZApi } from "../../providers/CENNZApiProvider";
+import Switch from "../../components/bridge/Switch";
+import Deposit from "../../components/bridge/Deposit";
+import Withdraw from "../../components/bridge/Withdraw";
 
-const ETH_CHAIN_ID = process.env.NEXT_PUBLIC_ETH_CHAIN_ID;
-
-const Home: React.FC<{}> = () => {
-	const [bridgeState, setBridgeState] = useState("");
-	const { selectedAccount } = useWallet();
+const Emery: React.FC<{}> = () => {
+	const [isDeposit, toggleIsDeposit] = useState<boolean>(true);
+	const { initBlockchain, Account } = useBlockchain();
 	const { api, initApi } = useCENNZApi();
 
 	useEffect(() => {
@@ -21,23 +21,67 @@ const Home: React.FC<{}> = () => {
 	useEffect(() => {
 		(async () => {
 			const { ethereum }: any = window;
-			const ethChainId = await ethereum.request({ method: "eth_chainId" });
-			const CENNZnetAccount = window.localStorage.getItem("CENNZnet-account");
+			const accounts = await ethereum.request({
+				method: "eth_requestAccounts",
+			});
 
-			((ethChainId === "0x1" && ETH_CHAIN_ID === "1") ||
-				(ethChainId === "0x2a" && ETH_CHAIN_ID === "42")) &&
-			CENNZnetAccount
-				? setBridgeState("emery")
-				: setBridgeState("connect");
+			if (!Account) initBlockchain(ethereum, accounts);
 		})();
-	}, [selectedAccount]);
+	}, [Account, initBlockchain]);
 
 	return (
-		<BlockchainProvider>
-			{bridgeState === "emery" && <Emery />}
-			{bridgeState === "connect" && <Connect setBridgeState={setBridgeState} />}
-		</BlockchainProvider>
+		<>
+			<Typography
+				sx={{
+					position: "absolute",
+					top: "4.5%",
+					left: "16%",
+					fontFamily: "Teko",
+					fontStyle: "normal",
+					fontWeight: "bold",
+					fontSize: "24px",
+					lineHeight: "124%",
+					color: "black",
+					letterSpacing: "1px",
+				}}
+			>
+				ETHEREUM BRIDGE
+			</Typography>
+			<Switch isDeposit={isDeposit} toggleIsDeposit={toggleIsDeposit} />
+			{Account && (
+				<Frame
+					sx={{
+						top: "12%",
+						right: "5%",
+						backgroundColor: "#FFFFFF",
+						cursor: "copy",
+					}}
+					onClick={() => navigator.clipboard.writeText(Account)}
+				>
+					<>
+						<Heading
+							sx={{
+								color: "primary.main",
+								ml: "10px",
+								mt: "3px",
+								fontSize: "20px",
+								flexGrow: 1,
+							}}
+						>
+							METAMASK
+						</Heading>
+						<SmallText sx={{ color: "black", fontSize: "16px" }}>
+							{Account.substring(0, 6).concat(
+								"...",
+								Account.substring(Account.length - 4, Account.length)
+							)}
+						</SmallText>
+					</>
+				</Frame>
+			)}
+			{isDeposit ? <Deposit /> : <Withdraw />}
+		</>
 	);
 };
 
-export default Home;
+export default Emery;
